+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - **Breaking:** Remove `unallocated`, becomes `new` or `new_in`
+- **Breaking:** Remove `into_aligned`, use `with_settings`
+- **Breaking:** Remove `as_mut_aligned`, use `borrow_mut_with_settings`
 - **Breaking:** Remove `alloc_layout` (`allocate_layout` provides the same functionality)
 - **Breaking:** Remove `BaseAllocator`
 - **Breaking:** Remove deprecated api